requirements.txt
numpy
faiss-cpu
sentence-transformers
pymupdf
beautifulsoup4
lxml
tqdm
orjson